+++ web/web.go
@@ -392,6 +392,7 @@ func New(logger *slog.Logger, o *Options) *Handler {
 		o.ConvertOTLPDelta,
 		o.NativeOTLPDeltaIngestion,
 		o.CTZeroIngestionEnabled,
+		o.LookbackDelta,
 	)
 
 	if o.RoutePrefix != "/" {